Address feedback, implement CommandChannel
